Move webpush to notifications tab
