+++ app/javascript/packs/forkMonitorApp/components/nodes.jsx
@@ -28,63 +28,6 @@ class Nodes extends React.Component {
     this.getChaintips = this.getChaintips.bind(this);
     this.getNodes = this.getNodes.bind(this);
     this.getInvalidBlocks = this.getInvalidBlocks.bind(this);
-
-    function toByteArray(hexString) {
-      var result = [];
-      for (var i = 0; i < hexString.length; i += 2) {
-        result.push(parseInt(hexString.substr(i, 2), 16));
-      }
-      return new Uint8Array(result);
-    }
-
-    var vapidPublicKey = toByteArray(process.env['VAPID_PUBLIC_KEY']);
-
-    function checkNotifs(obj){
-      if (!("Notification" in window)) {
-        return;
-      }
-      // Check whether notification permissions have already been granted
-      if (Notification.permission === "granted") {
-        getKeys();
-        return;
-      }
-      // TODO: move this to button
-      // Ask the user for permission
-      if (Notification.permission !== 'denied') {
-        Notification.requestPermission(function (permission) {
-          // If the user accepts, let's create a notification
-          if (permission === "granted") {
-            getKeys();
-          }
-        });
-      }
-   }
-
-   function getKeys(){
-     navigator.serviceWorker.register('/serviceworker.js', {scope: './'})
-       .then(function(registration) {
-         return registration.pushManager.getSubscription()
-           .then(function(subscription) {
-             if (subscription) {
-               return subscription;
-             }
-             return registration.pushManager.subscribe({
-               userVisibleOnly: true,
-               applicationServerKey: vapidPublicKey
-             }).then(function(subscription) {
-               axios.post('/api/v1/subscriptions', {
-                 subscription: subscription
-               }).then(function (response) {
-                 return response.data;
-               }).catch(function (error) {
-                 console.error(error);
-               });
-             });
-           });
-       });
-     }
-
-     checkNotifs()
   }
 
   componentDidMount() {
